Module espresso_core::kv_merkle_tree

Structs

KVMerkleProof

Structure for proofs in the KVMT

Enums

KVMerkleTerminalNode

Terminal nodes in a KVMT are either a leaf or an empty subtree

KVMerkleTree

The core enum of the key-value Merkle tree. Making use of KVTreeHash, it is generic over branch-arity (number of branches). Note: this type implements PartialEq so that containing types can derive PartialEq, mostly for testing purposes. The implementation tests for logical equality of the represented set, ignoring sparseness. That is, any two sets with the same root hash will compare equal, even if the elements retained in memory are different between the two sets.

Functions

kv_merkle_lw_multi_insert