Module espresso_core::merkle_tree

Implementation of the Merkle tree data structure.

At a high level the Merkle tree is a ternary tree and the hash function H used is SHA3_256 function. The node values are 32 bytes and each internal node value is obtained by computing v:=H(1u8,a,b,c) where a,b,c are the values of the left,middle and right child respectively. Leaf values for an element (uid,elem) is obtained as H(0u8,little_endian(uid),CanonicalSerialize(elem)). The tree height is fixed during initial instantiation and a new leaf will be inserted at the leftmost available slot in the tree.

Structs

FilledMTBuilder

MerkleCommitment

A wrapper of the merkle root, together with the tree information.

MerkleLeaf

Data struct for a merkle leaf.

MerkleLeafProof

Inclusive proof of a merkle leaf.

MerklePath

An authentication path of a ternary Merkle tree. While node information can come in any order, in this implementation we expect the first item to correspond to the leaf and the last to the root.

MerklePathNode

A 3-ary Merkle tree node.

MerkleTree

Data struct of a merkle tree.

NodeValue

Represents the value for a node in the merkle tree.

Enums

LookupResult

The result of querying at an index in the tree

MerkleFrontier

A wrapper of the merkle membership proof.

MerkleNode 🔒

Data structure storing the information of a node in the Merkle tree. The node has at most three children.

NodePos

Enum for identifying a position of a node (left, middle or right).

Constants

DOM_SEP_INT_NODE

DOM_SEP_LEAF_NODE

NODE_VALUE_LEN

Functions

hash 🔒

Hash function used to compute an internal node value