Module espresso_core::reward

Modules

canonical

eligibility

Structs

CollectRewardBody

Reward Collection Transaction Note

CollectRewardNote

Reward Collection Transaction Note

CollectedRewards

Previously collected rewards are recorded in (StakingKey, view_number) pairs

CollectedRewardsHistory

Sliding window for reward collection

CollectedRewardsTag

Identifying tag for CollectedReward

EligibilityWitness

EmptyStakeTableCommitmentSetSnafu 🔒

SNAFU context selector for the RewardError::EmptyStakeTableCommitmentSet variant

InvalidViewNumberSnafu 🔒

SNAFU context selector for the RewardError::InvalidViewNumber variant

KeyNotEligibleSnafu 🔒

SNAFU context selector for the RewardError::KeyNotEligible variant

ProofNotInMemorySnafu 🔒

SNAFU context selector for the RewardError::ProofNotInMemory variant

RewardAlreadyCollectedSnafu 🔒

SNAFU context selector for the RewardError::RewardAlreadyCollected variant

RewardNoteProofs

Auxiliary info and proof for CollectRewardNote

SerializationSnafu 🔒

SNAFU context selector for the RewardError::SerializationError variant

SignatureSnafu 🔒

SNAFU context selector for the RewardError::SignatureError variant

StakingKeyNotFoundSnafu 🔒

SNAFU context selector for the RewardError::StakingKeyNotFound variant

VrfProof

Proof for Vrf output

Enums

RewardError

Reward Transaction Errors.

Functions

compute_reward_amount

Compute the allowed stake amount given committee size, view_number and number of votes Hard-coded to 0 for FST

Type Definitions

CollectedRewardsDigest

CollectedRewardsHash

Hash for set Merkle tree for all of the previously-collected rewards

CollectedRewardsProof

CollectedRewardsProofs

CollectedRewards proofs, organized by the root hash for which they are valid.

CollectedRewardsSet