Enum espresso_core::state::ValidationError

pub enum ValidationError {

    NullifierAlreadyExists {
        nullifier: Nullifier,
    },
    BadNullifierProof {},
    MissingNullifierProof {},
    ConflictingNullifiers {},
    Failed {},
    BadMerkleLength {},
    BadMerkleLeaf {},
    BadMerkleRoot {},
    BadMerklePath {},
    CryptoError {
        err: Result<TxnApiError, String>,
    },
    UnsupportedTransferSize {
        num_inputs: usize,
        num_outputs: usize,
    },
    UnsupportedFreezeSize {
        num_inputs: usize,
    },
    InconsistentHelperProofs,
    UnexpectedGenesis,
    IncorrectParent,
    InvalidTime,
    BadCollectRewardNote,
    RewardAlreadyCollected {
        reward: CollectedRewards,
    },
    BadCollectedRewardProof {},
    RewardAmountTooLarge,
    BadStakeTableProof {},
    BadStakeTableCommitmentsProof {},
    BadFeeCalculation {},

}

Validation errors.

Variants

NullifierAlreadyExists

Fields

nullifier: Nullifier

A record was already spent.

BadNullifierProof

Fields

An invalid nullifier proof.

MissingNullifierProof

Fields

ConflictingNullifiers

Fields

The transaction being added to a block contains a nullifier already present in another transaction in the block.

Failed

Fields

A generic failure.

BadMerkleLength

Fields

An incorrect Merkle path length.

BadMerkleLeaf

Fields

An invalid Merkle leaf.

BadMerkleRoot

Fields

An incorrect Merkle root.

BadMerklePath

Fields

An invalid Merkle path.

CryptoError

Fields

err: Result<TxnApiError, String>

An error from the Jellyfish library

Note: This is used to wrap [TxnApiError] because it cannot be serialized. TxnApiError cannot be serialized because it depends on many foreign error types which do not implement the Serialize trait. Instead, if we have to serialize this variant, we will serialize Ok(err) to Err(format(err)), and when we deserialize we will at least preserve the variant CryptoError and a String representation of the underlying error.

UnsupportedTransferSize

Fields

num_inputs: usize

num_outputs: usize

The transfer transaction has an unsupported number of inputs or outputs.

Note: For transactions with fewer inputs or outputs than supported, the transaction should be padded with dummy transactions. If transactions with a greater number of inputs or outputs are required, then the universal parameter set for the ledger must be updated.

UnsupportedFreezeSize

Fields

num_inputs: usize

The freeze transaction has an unsupported number of inputs or outputs.

Note: For transactions with fewer inputs or outputs than supported, the transaction should be padded with dummy transactions. If transactions with a greater number of inputs or outputs are required, then the universal parameter set for the ledger must be updated.

InconsistentHelperProofs

Block transaction order doesn’t match helper proofs

UnexpectedGenesis

A genesis transaction was included in a non-genesis block

IncorrectParent

Attempted to apply a block to a state which was not its intended parent state

InvalidTime

Attempted to apply a block with a time in the past

BadCollectRewardNote

Bad CollectRewardNote

RewardAlreadyCollected

Fields

reward: CollectedRewards

A record was already spent.

BadCollectedRewardProof

Fields

An invalid Collected Reward proof.

RewardAmountTooLarge

Stake amount in transaction does not match amount in stake table

BadStakeTableProof

Fields

verification error for stake table proof

BadStakeTableCommitmentsProof

Fields

verification error for stake table commitments proof

BadFeeCalculation

Fields

Error when calculating block fees