Struct espresso_core::state::ValidatorState

pub struct ValidatorState {

    pub chain: ChainVariables,
    pub prev_commit_time: u64,
    pub block_height: u64,
    pub prev_state: Option<LedgerStateCommitment>,
    pub record_merkle_commitment: MerkleCommitment,
    pub record_merkle_frontier: MerkleFrontier,
    pub past_record_merkle_roots: RecordMerkleHistory,
    pub past_nullifiers: NullifierHistory,
    pub prev_block: BlockCommitment,
    pub stake_table: StakeTableMap,
    pub total_stake: Amount,
    pub stake_table_commitments: StakeTableCommFrontier,
    pub stake_table_commitments_commitment: StakeTableCommCommitment,
    pub collected_rewards: KVMerkleTree<CollectedRewardsHash>,

}

The working state of the ledger

Only the previous state is represented as a commitment. Other values are present in full.

Fields
chain: ChainVariablesprev_commit_time: u64

The consensus time at which this state was created.

“Consensus time” is an opaque notion of time which is meaningful in the consensus layer. From outside the consensus protocol, it can be treated as a monotonically (but possibly non-consecutively) increasing counter.

block_height: u64

The number of blocks in the chain which led to this state.

This field can also be used to determine the index of the block which created this state or the next block to be appended to this state. The 0-based index of a block appended to a chain of n blocks is n, so block_height is the index of the next block to be appended, and block_height - 1 is the index of the previous block, which created this state. (The default, pre-genesis state has block_height == 0, since it was not created by any block.)

prev_state: Option<LedgerStateCommitment>record_merkle_commitment: MerkleCommitment

The current record Merkle commitment

record_merkle_frontier: MerkleFrontier

The current frontier of the record Merkle tree

past_record_merkle_roots: RecordMerkleHistory

A list of recent record Merkle root hashes for validating slightly out-of-date transactions

past_nullifiers: NullifierHistory

Nullifiers from recent blocks, which allows validating slightly out-of-date-transactions

prev_block: BlockCommitmentstake_table: StakeTableMap

Staking table. For fixed-stake, this will be the same each round

total_stake: Amount

Total amount staked for the current table

stake_table_commitments: StakeTableCommFrontier

Keeps track of previous stake tables and their total stake

stake_table_commitments_commitment: StakeTableCommCommitment

Commitment to stake table commitments set

collected_rewards: KVMerkleTree<CollectedRewardsHash>

Track already-collected rewards via (staking_key, block number) tuples

Implementations

impl ValidatorState

pub const HISTORY_SIZE: usize = 10usize

The number of recent record Merkle tree root hashes the validator should remember

Transactions can be validated without resubmitting or regenerating the ZKPs as long as they were generated using a validator state that is in the last HISTORY_SIZE states.

pub fn new(
chain: ChainVariables,
record_merkle_frontier: MerkleTree,
stake_table_map: StakeTableMap,
total_stake: Amount,
stake_table_commitments_mt: StakeTableCommMT
) -> Self

pub fn commit(&self) -> LedgerStateCommitment

Cryptographic commitment to the validator state

pub fn nullifiers_root(&self) -> Hash

pub fn validate_block_check(
&self,
now: u64,
txns: Block,
txns_helper_proofs: Vec<EspressoTxnHelperProofs>
) -> Result<(Block, NullifierProofs, VerifiedRewards), ValidationError>

Validate a block of elaborated transactions

Checks the following

None of the nullifiers are used more than once

Keys are available for the numbers of transaction inputs and outputs

The Merkle roots are recent enough

The zero knowledge proofs in each of the transactions verifies

If valid, return the input transactions and proofs, otherwise return a validation error. A future implementation may return only the valid transactions and proofs.

Errors

ValidationError::BadMerkleRoot

ValidationError::BadNullifierProof

ValidationError::CryptoError

ValidationError::NullifierAlreadyExists

ValidationError::UnsupportedFreezeSize

ValidationError::UnsupportedTransferSize

ValidationError::PreviouslyCollectedReward

ValidationError::RewardAmountTooLarge

pub fn validate_and_apply(
&mut self,
now: u64,
txns: Block,
proofs: Vec<EspressoTxnHelperProofs>
) -> Result<ValidationOutputs, ValidationError>

Performs validation for a block, updating the ValidatorState.

If successful, returns

the UIDs of the newly created records

updated nullifier non-membership proofs for all of the nullifiers in txns, relative to the nullifier set at the time this function was invoked, in the form of a sparse reperesentation of a SetMerkleTree

Errors

ValidationError::BadNullifierProof

ValidationError::BadMerklePath

Panics

Panics if the record Merkle commitment is inconsistent with the record Merkle frontier.

pub fn update_nullifier_proofs(
&self,
txns: &[EspressoTransaction],
proofs: Vec<EspressoTxnHelperProofs>
) -> Result<SetMerkleTree, ValidationError>

pub fn update_records_frontier(&self, txns: &[EspressoTransaction]) -> MerkleTree