Trait espresso_core::tree_hash::KVTreeHash

pub trait KVTreeHash: Clone + PartialEq + Eq + Debug {
    type Digest: Hash + Debug + Eq + PartialEq + Clone + Copy + Hash + CanonicalSerialize + CanonicalDeserialize;
    type Key: Debug + Clone + PartialEq + Eq + Hash + CanonicalSerialize + CanonicalDeserialize;
    type Value: Debug + Clone + PartialEq + Eq + Hash + CanonicalSerialize + CanonicalDeserialize;
    type BranchArityMinus1: AddLength<Self::Digest, U1> + ArrayLength<Self::Digest> + Debug + PartialEq;
    type MaxDepth: ArrayLength<u8> + Debug + PartialEq;

    fn empty_digest() -> Self::Digest;
    fn traversal_of_digest(d: Self::Digest) -> GenericArray<u8, Self::MaxDepth>;
    fn digest_of_traversal(
        t: &GenericArray<u8, Self::MaxDepth>
    ) -> Option<Self::Digest>;
    fn hash_leaf(key: Self::Key, value: Self::Value) -> Self::Digest;
    fn hash_key(key: Self::Key) -> Self::Digest;
    fn hash_branch(
        children: &GenericArray<Self::Digest, <Self::BranchArityMinus1 as AddLength<Self::Digest, U1>>::Output>
    ) -> Self::Digest;
}

A hash function usable for sparse merkle tree implementations.

Inherits several other traits for #[derive] ergonomics

Required Associated Types

type Digest: Hash + Debug + Eq + PartialEq + Clone + Copy + Hash + CanonicalSerialize + CanonicalDeserialize

The output of the hash function

type Key: Debug + Clone + PartialEq + Eq + Hash + CanonicalSerialize + CanonicalDeserialize

A data type for keys A data type for values

type Value: Debug + Clone + PartialEq + Eq + Hash + CanonicalSerialize + CanonicalDeserialize

type BranchArityMinus1: AddLength<Self::Digest, U1> + ArrayLength<Self::Digest> + Debug + PartialEq

The number of children each branch has minus one. This is due to the proof paths having sibling path lengths of n-1 and the generic implementation relying on type systems

type MaxDepth: ArrayLength<u8> + Debug + PartialEq

How many base-BranchArityMinus1 place values are in a Digest

Required Methods

fn empty_digest() -> Self::Digest

A unique digest for empty subtrees

fn traversal_of_digest(d: Self::Digest) -> GenericArray<u8, Self::MaxDepth>

Convert a digest to a traversal

fn digest_of_traversal(
t: &GenericArray<u8, Self::MaxDepth>
) -> Option<Self::Digest>

Convert a traversal back to a digest. Primarily relevant for testing. Should return None if and only if a traversal is not representable by a Digest.

fn hash_leaf(key: Self::Key, value: Self::Value) -> Self::Digest

Hash a leaf’s contents

fn hash_key(key: Self::Key) -> Self::Digest

Hash a key’s contents

fn hash_branch(
children: &GenericArray<Self::Digest, <Self::BranchArityMinus1 as AddLength<Self::Digest, U1>>::Output>
) -> Self::Digest

Hash a branch’s children: the number of siblings in a branch is one greater than the AddLengthSelf::BranchArityMinus1,U1